requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web app
streamlit>=1.28.0

# Progress bars
tqdm>=4.65.0

# Data science utilities
scikit-learn>=1.3.0

# Optional: For statistical analysis
statsmodels>=0.14.0
